h265: CB2 play codec h265 video fail in some board. [1/1]
PD#SWPL-83948

Problem:
Since the information of ucode version management cannot be obtained on
Linux, the swap mechanism is not used, so that problems will occur when
the ucode space is not enough.

Solution:
Get the version information of the ucode directly, and enable the swap
function when certain conditions are met.

Verify:
oppen

Change-Id: I953605d484d2be5f384bd49a0cb4be90c148756d
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>